Update sfzh.py to prevent numpy >= 1.24.0 issue
replace np.float with float to reflect changes made to numpy from version 1.24.0 update

+++ synthesizer/parametric/sfzh.py
@@ -195,7 +195,7 @@ class SFH:
     class Common:
 
         def sfr(self, age):
-            if type(age) == np.float:
+            if type(age) == float:
                 return self.sfr_(age)
             elif type(age) == np.ndarray:
                 return np.array([self.sfr_(a) for a in age])
